requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
streamlit==1.32.0
python-dotenv==1.0.1
groq==0.9.0
chromadb==0.5.0
sentence-transformers==2.7.0
langchain==0.1.20
pdfplumber==0.10.3
beautifulsoup4==4.12.2
python-multipart==0.0.6
requests==2.31.0
pydantic==2.1.1
httpx==0.25.2
